# Implement Managed Simulation Server for MCP