… development/managed_simulation_server_plan.md
… examples/frozen_lake_mcp_complete/mcp_server/managed_simulation_server.py
… examples/frozen_lake_mcp_complete/tests/recordings/managed_simulation_trajectory.jsonl
… examples/frozen_lake_mcp_complete/tests/recordings/production_trajectory.jsonl
… examples/frozen_lake_mcp_complete/tests/recordings/simulation_trajectory.jsonl
… examples/frozen_lake_mcp_complete/tests/test_record_and_replay_e2e.py